requirements.txt
openenv>=0.1.13
openenv-core>=0.1.13
fastapi>=0.115.0
uvicorn[standard]>=0.24.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0